feat: extract importer.go, extend GetDealStats
