+++ db.go
@@ -21,6 +21,7 @@ const dbSchema = `
 		comm_p VARCHAR(255) NOT NULL,
 		state VARCHAR(255) NOT NULL,
 		mode VARCHAR(255) NOT NULL,
+		size BIGINT,
 		message TEXT,
 		created_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 	);
@@ -32,6 +33,7 @@ type DbImportedDeal struct {
 	CommP       string `json:"comm_p"`
 	State       string `json:"state"`
 	Mode        Mode   `json:"mode"`
+	Size        int64  `json:"size"`
 	Message     string `json:"message"`
 	CreatedDate string `json:"created_date"`
 }
@@ -59,14 +61,14 @@ func OpenDiDB(root string) (*diDB, error) {
 	}, nil
 }
 
-func (d *diDB) InsertDeal(dealUuid string, commP string, success bool, mode Mode) error {
+func (d *diDB) InsertDeal(dealUuid string, commP string, success bool, mode Mode, message string, size int64) error {
 	var state string
 	if success {
 		state = PENDING
 	} else {
 		state = FAILURE
 	}
-	_, err := d.db.Exec("INSERT INTO imported_deals (deal_uuid, comm_p, state, mode) VALUES (?, ?, ?, ?)", dealUuid, commP, state, mode)
+	_, err := d.db.Exec("INSERT INTO imported_deals (deal_uuid, comm_p, state, mode, message, size) VALUES (?, ?, ?, ?, ?, ?)", dealUuid, commP, state, mode, message, size)
 
 	if err != nil {
 		return fmt.Errorf("insert deal: %w", err)
@@ -75,14 +77,38 @@ func (d *diDB) InsertDeal(dealUuid string, commP string, success bool, mode Mode
 }
 
 type DealStats struct {
-	TotalImported int
+	TotalImported Stat
+	Pending       Stat
+	Success       Stat
+	Failed        Stat
+}
+
+type Stat struct {
+	Count uint
+	Bytes uint
 }
 
 func (d *diDB) GetDealStats() (DealStats, error) {
 	var stats DealStats
-	err := d.db.QueryRow("SELECT COUNT(*) FROM imported_deals").Scan(&stats.TotalImported)
+	err := d.db.QueryRow("SELECT COUNT(*), SUM(size) FROM imported_deals").Scan(&stats.TotalImported.Count, &stats.TotalImported.Bytes)
 	if err != nil {
-		return DealStats{}, fmt.Errorf("get deal stats: %w", err)
+		return stats, fmt.Errorf("get total deal stats: %w", err)
 	}
+
+	err = d.db.QueryRow("SELECT COUNT(*), SUM(size) FROM imported_deals WHERE state = 'PENDING'").Scan(&stats.Pending.Count, &stats.Pending.Bytes)
+	if err != nil {
+		return stats, fmt.Errorf("get pending deal stats: %w", err)
+	}
+
+	err = d.db.QueryRow("SELECT COUNT(*), SUM(size) FROM imported_deals WHERE state = 'SUCCESS'").Scan(&stats.Success.Count, &stats.Success.Bytes)
+	if err != nil {
+		return stats, fmt.Errorf("get success deal stats: %w", err)
+	}
+
+	err = d.db.QueryRow("SELECT COUNT(*), SUM(size) FROM imported_deals WHERE state = 'FAILED'").Scan(&stats.Failed.Count, &stats.Failed.Bytes)
+	if err != nil {
+		return stats, fmt.Errorf("get failed deal stats: %w", err)
+	}
+
 	return stats, nil
 }
